Refactor README y props GotenPaginator
