+++ example/src/components/List.js
@@ -1,30 +1,24 @@
-import React, { Component } from 'react';
+import React from 'react'
 
-class List extends Component {
 
-    getElements = () =>{
-        return (
-            <ul>
-                {
-                    this.props.elements.map((element, i) => {
-                    return (
-                        <li key={i}>
-                            {element}
-                        </li>
-                        )
-                    })
-                }
-            </ul>
-        )
-    }
+export default class List extends React.Component {
+    getElements = () => (
+        <ul>
+            {this.props.elements.map((element, i) => {
+                return (
+                    <li key={i}>
+                        {element}
+                    </li>
+                )
+            })}
+        </ul>
+    )
     
     render() {
         return (
-        <div>
-            { this.getElements() }
-        </div>
-        );
+            <div>
+                {this.getElements()}
+            </div>
+        )
     }
 }
-
-export default List;
